+++ test-support/helpers/ember-basic-dropdown.js
@@ -1,5 +1,0 @@
-import { click } from 'ember-native-dom-helpers';
-import registerHelpers from 'ember-basic-dropdown/test-support/helpers';
-export default registerHelpers;
-export { nativeTap, clickTrigger, tapTrigger, fireKeydown } from 'ember-basic-dropdown/test-support/helpers';
-export const nativeClick = click;